+++ easyarmorstands-plugin/src/main/java/me/m56738/easyarmorstands/dialog/DialogBodyEntry.java
@@ -1,0 +1,11 @@
+package me.m56738.easyarmorstands.dialog;
+
+import io.papermc.paper.registry.data.dialog.body.DialogBody;
+
+import java.util.List;
+import java.util.Locale;
+
+@SuppressWarnings("UnstableApiUsage")
+public interface DialogBodyEntry {
+    void populateBody(List<DialogBody> body, Locale locale);
+}