+++ easyarmorstands-plugin/src/main/java/me/m56738/easyarmorstands/element/ElementMenuListener.java
@@ -36,6 +36,7 @@ import me.m56738.easyarmorstands.property.button.BlockDataHandler;
 import me.m56738.easyarmorstands.property.button.BooleanHandler;
 import me.m56738.easyarmorstands.property.button.ButtonHandler;
 import me.m56738.easyarmorstands.property.button.ComponentHandler;
+import me.m56738.easyarmorstands.property.button.CustomNameHandler;
 import me.m56738.easyarmorstands.property.button.EnumHandler;
 import me.m56738.easyarmorstands.property.button.GravityHandler;
 import me.m56738.easyarmorstands.property.button.OptionalComponentHandler;
@@ -56,6 +57,7 @@ import org.bukkit.inventory.EquipmentSlot;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.MainHand;
 
+import java.util.Optional;
 import java.util.function.Function;
 
 public class ElementMenuListener implements Listener {
@@ -84,13 +86,21 @@ public class ElementMenuListener implements Listener {
         populator.addButton(BlockDisplayPropertyTypes.BLOCK, BlockDataHandler::new, Material.OAK_STAIRS, Component.translatable("easyarmorstands.property.block-display.block.pick"));
         populator.addButton(DisplayPropertyTypes.BILLBOARD, EnumHandler.provider(Display.Billboard.class), Material.IRON_BARS, Component.translatable("easyarmorstands.property.display.billboard.description"));
         populator.addButton(EntityPropertyTypes.AI, Material.REDSTONE_BLOCK, Component.translatable("easyarmorstands.property.ai.description"));
-        populator.addButton(EntityPropertyTypes.CUSTOM_NAME, OptionalComponentHandler.provider("/eas name set"), Material.NAME_TAG);
+        populator.addButton(EntityPropertyTypes.CUSTOM_NAME, property -> {
+            Property<Boolean> customNameVisibleProperty = container.getOrNull(EntityPropertyTypes.CUSTOM_NAME_VISIBLE);
+            ButtonHandler handler = customNameVisibleProperty != null
+                    ? new CustomNameHandler(property, customNameVisibleProperty)
+                    : new OptionalComponentHandler(property);
+            return new PropertyButtonBuilder<>(property)
+                    .icon(MenuIcon.of(Material.NAME_TAG))
+                    .handler(handler);
+        });
         populator.addButton(EntityPropertyTypes.GLOWING, Material.GLOWSTONE_DUST, Component.translatable("easyarmorstands.property.glow.description"));
         populator.addButton(EntityPropertyTypes.SILENT, Material.NOTE_BLOCK, Component.translatable("easyarmorstands.property.silent.description"));
         populator.addButton(EntityPropertyTypes.VISIBLE, Material.POTION);
         populator.addButton(InteractionPropertyTypes.RESPONSIVE, Material.OAK_PRESSURE_PLATE, Component.translatable("easyarmorstands.property.interaction.responsive.description"));
         populator.addButton(ItemDisplayPropertyTypes.TRANSFORM, EnumHandler.provider(ItemDisplay.ItemDisplayTransform.class), Material.STICK, Component.translatable("easyarmorstands.property.item-display.transform.description"));
-        populator.addButton(MannequinPropertyTypes.DESCRIPTION, OptionalComponentHandler.provider("/eas description set"), Material.NAME_TAG, Component.translatable("easyarmorstands.property.mannequin.description.description"));
+        populator.addButton(MannequinPropertyTypes.DESCRIPTION, (Function<Property<Optional<Component>>, OptionalComponentHandler>) OptionalComponentHandler::new, Material.NAME_TAG, Component.translatable("easyarmorstands.property.mannequin.description.description"));
         populator.addButton(MannequinPropertyTypes.IMMOVABLE, Material.BEDROCK, Component.translatable("easyarmorstands.property.mannequin.immovable.description"));
         populator.addButton(MannequinPropertyTypes.MAIN_HAND, EnumHandler.provider(MainHand.class), Material.IRON_SWORD);
         populator.addButton(MannequinPropertyTypes.SKIN_PART_VISIBLE.get(SkinPart.CAPE), Material.BLUE_BANNER);
@@ -106,7 +116,7 @@ public class ElementMenuListener implements Listener {
         populator.addButton(TextDisplayPropertyTypes.BACKGROUND, TextBackgroundHandler::new, Material.STONE_SLAB);
         populator.addButton(TextDisplayPropertyTypes.SEE_THROUGH, Material.GLOWSTONE_DUST);
         populator.addButton(TextDisplayPropertyTypes.SHADOW, Material.STONE);
-        populator.addButton(TextDisplayPropertyTypes.TEXT, ComponentHandler.provider("/eas text set"), Material.NAME_TAG);
+        populator.addButton(TextDisplayPropertyTypes.TEXT, ComponentHandler::new, Material.NAME_TAG);
 
         for (EquipmentSlot slot : EquipmentSlot.values()) {
             Property<ItemStack> property = container.getOrNull(EntityPropertyTypes.EQUIPMENT.get(slot));