+++ easyarmorstands-plugin/src/main/java/me/m56738/easyarmorstands/property/button/CustomNameHandler.java
@@ -4,16 +4,21 @@ import me.m56738.easyarmorstands.EasyArmorStandsPlugin;
 import me.m56738.easyarmorstands.api.menu.click.MenuClickContext;
 import me.m56738.easyarmorstands.api.property.Property;
 import me.m56738.easyarmorstands.clipboard.Clipboard;
+import me.m56738.easyarmorstands.dialog.BooleanDialogEntry;
 import me.m56738.easyarmorstands.dialog.DialogBuilder;
-import me.m56738.easyarmorstands.dialog.TextDialogEntry;
+import me.m56738.easyarmorstands.dialog.OptionalTextDialogEntry;
 import net.kyori.adventure.text.Component;
 import org.bukkit.entity.Player;
 
-public class ComponentHandler implements ButtonHandler {
-    private final Property<Component> property;
+import java.util.Optional;
 
-    public ComponentHandler(Property<Component> property) {
+public class CustomNameHandler implements ButtonHandler {
+    private final Property<Optional<Component>> property;
+    private final Property<Boolean> customNameVisibleProperty;
+
+    public CustomNameHandler(Property<Optional<Component>> property, Property<Boolean> customNameVisibleProperty) {
         this.property = property;
+        this.customNameVisibleProperty = customNameVisibleProperty;
     }
 
     @Override
@@ -23,12 +28,14 @@ public class ComponentHandler implements ButtonHandler {
         if (context.isShiftClick()) {
             Clipboard clipboard = plugin.getClipboard(player);
             clipboard.handlePropertyShiftClick(property);
+            clipboard.handlePropertyShiftClick(customNameVisibleProperty);
         } else if (context.isLeftClick()) {
             context.closeMenu();
             plugin.getServer().getScheduler().runTask(plugin, () -> {
                 DialogBuilder builder = new DialogBuilder(player.locale());
                 builder.setTitle(property.getType().getName());
-                builder.addEntry(new TextDialogEntry("text", property));
+                builder.addEntry(new OptionalTextDialogEntry("text", property));
+                builder.addEntry(new BooleanDialogEntry("visible", customNameVisibleProperty));
                 player.showDialog(builder.build());
             });
         }