+++ easyarmorstands-plugin/src/main/java/me/m56738/easyarmorstands/property/button/ToggleHandler.java
@@ -3,6 +3,7 @@ package me.m56738.easyarmorstands.property.button;
 import me.m56738.easyarmorstands.EasyArmorStandsPlugin;
 import me.m56738.easyarmorstands.api.menu.click.MenuClickContext;
 import me.m56738.easyarmorstands.api.property.Property;
+import me.m56738.easyarmorstands.clipboard.Clipboard;
 import me.m56738.easyarmorstands.message.Message;
 
 public abstract class ToggleHandler<T> implements ButtonHandler {
@@ -16,8 +17,7 @@ public abstract class ToggleHandler<T> implements ButtonHandler {
     public void onClick(MenuClickContext context) {
         boolean changed;
         if (context.isShiftClick()) {
-            EasyArmorStandsPlugin.getInstance().getClipboard(context.player())
-                    .handlePropertyShiftClick(property);
+            handlePropertyShiftClick(EasyArmorStandsPlugin.getInstance().getClipboard(context.player()));
             return;
         } else if (context.isLeftClick()) {
             changed = toggleNextValue();
@@ -31,6 +31,10 @@ public abstract class ToggleHandler<T> implements ButtonHandler {
         } else {
             context.player().sendMessage(Message.error("easyarmorstands.error.cannot-change"));
         }
+    }
+
+    protected void handlePropertyShiftClick(Clipboard clipboard) {
+        clipboard.handlePropertyShiftClick(property);
     }
 
     protected abstract boolean toggleNextValue();